+++ app/services/normalizer/phdcomics_normalizer.rb
@@ -1,0 +1,9 @@
+module Normalizer
+  class PhdcomicsNormalizer < RssNormalizer
+    private
+
+    def normalize_content
+      strip_html(raw_data["title"])
+    end
+  end
+end